Develop

+++ internal/match_strategy.go
@@ -1,5 +1,5 @@
 package match
 
 type strategy interface {
-	matchStrings(s1 string, s2 string) float32
+	matchStrings(s1 string, s2 string) int
 }
